// webapp/src/app/core/api/v1/model/note.ts
/**
 * permaAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Note { 
    noteId?: number | null;
    categoryId?: number;
    categoryTitle?: string | null;
    subject?: string | null;
    status?: string | null;
    createdBy?: string | null;
    createdDate?: string;
    content?: string | null;
}